create datavase models

+++ migrations/sqls/20210623123923-orders-table-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
     status VARCHAR(100) NOT NULL,
-    user_id bigint REFERENCES users(id)
+    user_id integer REFERENCES users(id)
 );
